refactor(arrowmap): split AppendValue into duckdb-free core + driver hooks
Step 2 of the binary-split plan. Moves AppendValue out of duckdbservice
into arrowmap so callers that don't need duckdb-go's driver value types
(duckdb.Interval, Decimal, UUID, OrderedMap, Map) can use it without
pulling libduckdb into their import graph.

The split uses a registration hook rather than a build tag:

  - arrowmap.AppendValue handles all arrow-native and Go-native value
    types directly. It first consults any registered Appender hooks
    before falling back to the built-in switch.
  - duckdbservice/appender_init.go's init() registers a single hook
    that handles duckdb.Interval / Decimal / UUID / OrderedMap / Map.
  - When duckdbservice is linked into a binary (worker, standalone),
    the duckdb types get full coverage automatically. When it isn't
    (a future controlplane-only binary), the duckdb cases are dead
    code paths that wouldn't fire anyway because duckdb-go is the
    only thing producing those typed values.

OrderedMapValue moved from server/flight_executor.go into arrowmap so
the MAP branch in arrowmap.AppendValue can switch on it without
arrowmap depending on the server package. server.OrderedMapValue is
preserved as a type alias for backward compatibility with existing
call sites.

server/flightsqlingress no longer imports duckdbservice — it uses
arrowmap.AppendValue directly. The package still transitively pulls
duckdb-go via its `server` import; getting `server` itself duckdb-free
is the next chunk of work.

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go test -short ./duckdbservice/... ./server/flightsqlingress/...
    all green
  - go list -deps ./duckdbservice/arrowmap | grep duckdb-go is empty
    (arrowmap remains a pure leaf even with AppendValue + OrderedMapValue
    living in it)
